MismatchedDice component added
This is a first iteration of allowing dice objects to be combined, with
a shape for the logic that allows this to be reasonably easy, along
with fitting the previous structures as much as possible.

Additionally, there is a new component to accompany this change that is
specifically built for handling multiDice objects.